Add functionality to cancel rocket reservation from profile page
Style the profile page
