+++ src/pages/Profile/Profile.css
@@ -1,0 +1,23 @@
+.profile {
+  padding: 1rem 4rem;
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-start;
+}
+
+.profile__section-list {
+  list-style: none;
+  margin: 10px 0;
+}
+
+.profile__section-list li {
+  font-size: 18px;
+  line-height: 24px;
+  margin: 10px 2px;
+}
+
+.profile__section-list_container {
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+}
